Extract component logic to page
