+++ src/services/qbit/IProvider.ts
@@ -144,7 +144,7 @@ export default interface IProvider {
    * @param itemPath Feed path
    * @param articleId Article ID
    */
-  markAsRead(itemPath: string, articleId: string): Promise<void>
+  markAsRead(itemPath: string, articleId?: string): Promise<void>
 
   /**
    * Refresh a feed
